Your commit message

+++ src/main/java/net/engineeringdigest/journalApp/services/JournalEntryService.java
@@ -1,21 +1,42 @@
 package net.engineeringdigest.journalApp.services;
 
+import lombok.extern.slf4j.Slf4j;
 import net.engineeringdigest.journalApp.Entity.JournalEntry;
+import net.engineeringdigest.journalApp.Entity.User;
 import net.engineeringdigest.journalApp.repository.JournalEntryRepository;
 import org.bson.types.ObjectId;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
+import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Optional;
 
-@Component
+@Service
+@Slf4j
 public class JournalEntryService {
 
     @Autowired
     private JournalEntryRepository journalEntryRepository;
+    @Autowired
+    private UserService userService;
 
-    public void saveEntry(JournalEntry journalEntry) {
+
+    @Transactional
+    public void saveEntry(JournalEntry journalEntry, String username){
+        try{
+            User user = userService.findByUsername(username);
+            journalEntry.setDate(LocalDateTime.now());
+            JournalEntry saved = journalEntryRepository.save(journalEntry);
+            user.getJournalEntries().add(saved);
+            userService.saveUser(user);
+        } catch (Exception e) {
+            throw new RuntimeException("An error occurred while entry ",e);        }
+    }
+
+    public void saveEntry(JournalEntry journalEntry){
         journalEntryRepository.save(journalEntry);
     }
 
@@ -27,8 +48,21 @@ public class JournalEntryService {
         return journalEntryRepository.findById(id);
     }
 
-    public void deleteById(ObjectId id){
-        journalEntryRepository.deleteById(id);
+    @Transactional
+    public boolean deleteById(ObjectId id, String userName){
+        boolean removed = false;
+        try {
+            User user = userService.findByUsername(userName);
+            removed = user.getJournalEntries().removeIf(x -> x.getId().equals(id));
+            if (removed) {
+                journalEntryRepository.deleteById(id);
+                userService.saveUser(user);
+            }
+        } catch (Exception e) {
+            log.info("An error occurred while deleting journal entry ",e);
+            throw new RuntimeException("An error occurred while deleting ",e);
+        }
+        return removed;
     }
 
 }
